Show tags as completion list
When editing a note, it is possible to select an already exist tag from
a completion menu.

It needs the latest Electron because of an `input` `datalist` bug. See:
https://github.com/electron/electron/issues/9860

NOTE:
Clicking on tag does note selects it; needs pressing Enter or `blur`
event, because I cannot find the proper event to detect mouse click:
`onChange` and `onInput` is fired when selecting tag by keyboard and by
mouse too.